deploy dulu baru perbaiki lagi

+++ src/app/api/produk/[id]/route.ts
@@ -20,26 +20,44 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 // UPDATE produk
 export async function PUT(req: Request, { params }: { params: { id: string } }) {
   const id = Number(params.id);
-  const { nama, harga, stok, gambar_url } = await req.json();
+  const body = await req.json();
 
   try {
     const updated = await prisma.produk.update({
       where: { id },
-      data: { nama, harga, stok, gambar_url },
+      data: {
+        nama: body.nama,
+        harga: Number(body.harga),
+        stok: Number(body.stok),
+        description: body.description,
+      },
     });
 
     return NextResponse.json(updated);
-  } catch (error) {
-    console.error(error);
-    return NextResponse.json({ error: "Gagal update produk" }, { status: 500 });
+  } catch (err) {
+    console.error(err);
+    return NextResponse.json({ err: "Gagal update" }, { status: 500 });
   }
 }
+
 
 // DELETE produk
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
   const id = Number(params.id);
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID tidak valid" }, { status: 400 });
+  }
 
   try {
+    // Cek apakah produk digunakan dalam PesananDetail
+    const sedangDipakai = await prisma.pesananDetail.findFirst({
+      where: { produk_id: id },
+    });
+
+    if (sedangDipakai) {
+      return NextResponse.json({ error: "Produk tidak bisa dihapus karena sudah digunakan dalam pesanan." }, { status: 400 });
+    }
+
     await prisma.produk.delete({ where: { id } });
     return NextResponse.json({ message: "Produk berhasil dihapus" });
   } catch (error) {
@@ -47,3 +65,4 @@ export async function DELETE(req: Request, { params }: { params: { id: string } 
     return NextResponse.json({ error: "Gagal hapus produk" }, { status: 500 });
   }
 }
+
